Add FailureConverter concept
Fixes #142
