+++ temporalio/worker/activity.py
@@ -416,30 +416,25 @@ class _ActivityWorker:
                     temporalio.activity.logger.warning(
                         f"Completing as failure during heartbeat with error of type {type(err)}: {err}",
                     )
-                    await temporalio.exceptions.encode_exception_to_failure(
-                        err,
-                        self._data_converter,
-                        completion.result.failed.failure,
+                    await self._data_converter.encode_failure(
+                        err, completion.result.failed.failure
                     )
                 elif (
                     isinstance(err, asyncio.CancelledError)
                     and running_activity.cancelled_by_request
                 ):
                     temporalio.activity.logger.debug("Completing as cancelled")
-                    await temporalio.exceptions.encode_error_to_failure(
+                    await self._data_converter.encode_failure(
                         # TODO(cretz): Should use some other message?
                         temporalio.exceptions.CancelledError("Cancelled"),
-                        self._data_converter,
                         completion.result.cancelled.failure,
                     )
                 else:
                     temporalio.activity.logger.warning(
                         "Completing activity as failed", exc_info=True
                     )
-                    await temporalio.exceptions.encode_exception_to_failure(
-                        err,
-                        self._data_converter,
-                        completion.result.failed.failure,
+                    await self._data_converter.encode_failure(
+                        err, completion.result.failed.failure
                     )
             except Exception as inner_err:
                 temporalio.activity.logger.exception(
