+++ temporalio/worker/workflow_instance.py
@@ -104,6 +104,7 @@ class WorkflowInstanceDetails:
     """Immutable, serializable details for creating a workflow instance."""
 
     payload_converter_class: Type[temporalio.converter.PayloadConverter]
+    failure_converter_class: Type[temporalio.converter.FailureConverter]
     interceptor_classes: Sequence[Type[WorkflowInboundInterceptor]]
     defn: temporalio.workflow._Definition
     info: temporalio.workflow.Info
@@ -161,6 +162,7 @@ class _WorkflowInstanceImpl(
         WorkflowInstance.__init__(self)
         temporalio.workflow._Runtime.__init__(self)
         self._payload_converter = det.payload_converter_class()
+        self._failure_converter = det.failure_converter_class()
         self._defn = det.defn
         self._info = det.info
         self._extern_functions = det.extern_functions
@@ -292,7 +294,7 @@ class _WorkflowInstanceImpl(
             # Set completion failure
             self._current_completion.failed.failure.SetInParent()
             try:
-                temporalio.exceptions.apply_exception_to_failure(
+                self._failure_converter.to_failure(
                     err,
                     self._payload_converter,
                     self._current_completion.failed.failure,
@@ -403,7 +405,7 @@ class _WorkflowInstanceImpl(
                 command.respond_to_query.succeeded.response.CopyFrom(result_payloads[0])
             except Exception as err:
                 try:
-                    temporalio.exceptions.apply_exception_to_failure(
+                    self._failure_converter.to_failure(
                         err,
                         self._payload_converter,
                         command.respond_to_query.failed,
@@ -457,13 +459,13 @@ class _WorkflowInstanceImpl(
             handle._resolve_success(ret)
         elif job.result.HasField("failed"):
             handle._resolve_failure(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.result.failed.failure, self._payload_converter
                 )
             )
         elif job.result.HasField("cancelled"):
             handle._resolve_failure(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.result.cancelled.failure, self._payload_converter
                 )
             )
@@ -494,13 +496,13 @@ class _WorkflowInstanceImpl(
             handle._resolve_success(ret)
         elif job.result.HasField("failed"):
             handle._resolve_failure(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.result.failed.failure, self._payload_converter
                 )
             )
         elif job.result.HasField("cancelled"):
             handle._resolve_failure(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.result.cancelled.failure, self._payload_converter
                 )
             )
@@ -538,7 +540,7 @@ class _WorkflowInstanceImpl(
         elif job.HasField("cancelled"):
             self._pending_child_workflows.pop(job.seq)
             handle._resolve_failure(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.cancelled.failure, self._payload_converter
                 )
             )
@@ -557,7 +559,7 @@ class _WorkflowInstanceImpl(
         # We intentionally let this error if future is already done
         if job.HasField("failure"):
             fut.set_exception(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.failure, self._payload_converter
                 )
             )
@@ -576,7 +578,7 @@ class _WorkflowInstanceImpl(
         # We intentionally let this error if future is already done
         if job.HasField("failure"):
             fut.set_exception(
-                temporalio.exceptions.failure_to_error(
+                self._failure_converter.from_failure(
                     job.failure, self._payload_converter
                 )
             )
@@ -1222,7 +1224,7 @@ class _WorkflowInstanceImpl(
             command = self._add_command()
             command.fail_workflow_execution.failure.SetInParent()
             try:
-                temporalio.exceptions.apply_exception_to_failure(
+                self._failure_converter.to_failure(
                     err,
                     self._payload_converter,
                     command.fail_workflow_execution.failure,
@@ -1232,7 +1234,7 @@ class _WorkflowInstanceImpl(
         except asyncio.CancelledError as err:
             command = self._add_command()
             command.fail_workflow_execution.failure.SetInParent()
-            temporalio.exceptions.apply_exception_to_failure(
+            self._failure_converter.to_failure(
                 temporalio.exceptions.CancelledError(str(err)),
                 self._payload_converter,
                 command.fail_workflow_execution.failure,
@@ -1580,7 +1582,7 @@ class _ActivityHandle(temporalio.workflow.ActivityHandle[Any]):
         # We intentionally let this error if already done
         self._result_fut.set_result(result)
 
-    def _resolve_failure(self, err: Exception) -> None:
+    def _resolve_failure(self, err: BaseException) -> None:
         # If it was never started, we don't need to set this failure. In cases
         # where this is cancelled before started, setting this exception causes
         # a Python warning to be emitted because this future is never awaited
@@ -1725,7 +1727,7 @@ class _ChildWorkflowHandle(temporalio.workflow.ChildWorkflowHandle[Any, Any]):
         # We intentionally let this error if already done
         self._result_fut.set_result(result)
 
-    def _resolve_failure(self, err: Exception) -> None:
+    def _resolve_failure(self, err: BaseException) -> None:
         if self._start_fut.done():
             # We intentionally let this error if already done
             self._result_fut.set_exception(err)
